Move readme content into /README.rst
And use include directive in docs readme.rst.

This is done in order to prevent symlinks not being followed
on the mobile version of GitHub, which requires users to press
"show contents of README.rst" in order to see the entirity
of the readme. It does work for desktop, though, as desktop
users don't need to visit the README.rst file directly.